// Qualyt.Web/ClientApp/src/app/models/pivot-table-data.model.ts
export class PivotTableData {
  planName: string;
  laboratoryName: string;
  countryName: string;
  patientFullName: string;
  patientState: string;
  attentionAddress: string;
  attentionPlaceName: string;
  nurseFullName: string;
  doctorFullName: string;
  salesContactFullName: string;
  treatmentCode:string;
  treatmentState: string;
  productName:string;
  pathologyName:string;
  control:string;
}